requirements.txt
aniso8601==9.0.1
bcrypt==4.0.0
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
email-validator==1.2.1
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-RESTful==0.3.9
Flask-WTF==1.0.1
greenlet==1.1.3
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.2.0
psycopg2-binary==2.9.3
PyMySQL==1.0.2
pytz==2022.2.1
six==1.16.0
SQLAlchemy==1.4.40
Werkzeug==2.2.2
WTForms==3.0.1
